Testing out ingests of beta NLDAS-3
This code is pulling the beta netcdf files and will not work if deployed because of the long download time.  We could probably modify this to read from the netcdfs directly for each hour (instead of downloading the full daily file), but that should probably wait until the data is being generated operationally.